+++ lib/peon.js
@@ -4,11 +4,11 @@ const {
   watcher: { enabled: watcherEnabled, repositories: watcherRepositories },
   webhooks: { enabled: webhooksEnabled }
 } = require('./config')
-const { extractRepoName } = require('./utils')
-const Watcher = require('./watcher')
-const Dispatcher = require('./dispatcher')
-const webhooks = require('./webhooks')
-const status = require('./status')
+const Dispatcher = require('./build/dispatcher')
+const status = require('./status/status')
+const { extractRepoName } = require('./utils/misc')
+const Watcher = require('./watch/watcher')
+const webhooks = require('./watch/webhooks')
 
 module.exports = async function() {
   await status.init()
