+++ lib/status/render.js
@@ -2,20 +2,20 @@ const { dirname, resolve } = require('path')
 const { mkdir, readdir, readFile, writeFile } = require('fs-extra')
 const Handlebars = require('handlebars')
 
-const { statusDirectory, workingDirectory } = require('./config')
-const logger = require('./logger')
+const { statusDirectory, workingDirectory } = require('../config')
+const logger = require('../utils/logger')('status')
 
-const statusRoot = resolve(workingDirectory, 'status')
+const statusJSONDirectory = resolve(workingDirectory, 'status')
 
 Handlebars.registerHelper('date', function(timestamp) {
   return new Date(timestamp).toISOString()
 })
 
-const templateDir = resolve(dirname(__dirname), 'templates')
+const templateDir = resolve(dirname(dirname(__dirname)), 'templates')
 const templates = ['index', 'build']
 const compiled = {}
 
-async function renderStatus(now) {
+module.exports = async function renderStatus(now) {
   // Compile templates if not already done
   for (let t of templates) {
     if (!compiled[t]) {
@@ -28,7 +28,7 @@ async function renderStatus(now) {
 
   // Load last render info
   logger.debug('loading render info', { module: 'status/render' })
-  let renderInfoFile = resolve(statusRoot, 'peon-status.json')
+  let renderInfoFile = resolve(statusJSONDirectory, 'peon-status.json')
   let renderInfo
   try {
     renderInfo = JSON.parse(await readFile(renderInfoFile))
@@ -40,14 +40,14 @@ async function renderStatus(now) {
 
   // Load status for all repos
   let status = {}
-  for (let file of await readdir(statusRoot)) {
+  for (let file of await readdir(statusJSONDirectory)) {
     if (file === 'peon-status.json') {
       continue
     }
 
     logger.debug(`reading status file ${file}`, { module: 'status/render' })
     status[file.replace(/\.json$/, '')] = JSON.parse(
-      await readFile(resolve(statusRoot, file))
+      await readFile(resolve(statusJSONDirectory, file))
     )
   }
 
@@ -128,118 +128,5 @@ async function renderStatus(now) {
   renderInfo.lastRender = now
   await writeFile(renderInfoFile, JSON.stringify(renderInfo))
 
-  logger.info('updated status pages', { module: 'status/render' })
-}
-
-async function updateRepoStatus(repoName, updater) {
-  let ret, repoStatus
-  let now = Date.now()
-
-  let statusFile = resolve(statusRoot, `${repoName}.json`)
-
-  try {
-    await mkdir(statusRoot, { recursive: true })
-  } catch(e) {
-    if (e.code !== 'EEXIST') {
-      throw e
-    }
-  }
-
-  try {
-    repoStatus = JSON.parse(await readFile(statusFile))
-  } catch(e) {
-    repoStatus = {
-      nextBuildNum: 1,
-      builds: {}
-    }
-  }
-
-  ret = updater(repoStatus, now)
-
-  await writeFile(statusFile, JSON.stringify(repoStatus))
-  await renderStatus(now)
-
-  return ret
-}
-
-module.exports = {
-  async init() {
-    try {
-      await mkdir(resolve(statusRoot), { recursive: true })
-    } catch(e) {
-      if (e.code !== 'EEXIST') {
-        throw e
-      }
-    }
-
-    try {
-      await mkdir(resolve(statusDirectory, 'peon-status'), { recursive: true })
-    } catch(e) {
-      if (e.code !== 'EEXIST') {
-        throw e
-      }
-    }
-
-    await renderStatus(Date.now())
-  },
-
-  // Returns buildId
-  async startBuild(repoName, branch, sha) {
-    return await updateRepoStatus(repoName, (repoStatus, now) => {
-      let buildNum = repoStatus.nextBuildNum
-      repoStatus.nextBuildNum++
-      let buildId = `${repoName}#${buildNum}`
-
-      repoStatus.builds[buildId] = {
-        branch,
-        sha,
-        enqueued: now,
-        updated: now,
-        status: 'pending',
-        steps: []
-      }
-
-      return buildId
-    })
-  },
-
-  async updateBuildStep(buildId, description, status, output) {
-    let [repoName] = buildId.split('#')
-
-    await updateRepoStatus(repoName, (repoStatus, now) => {
-      let build = repoStatus.builds[buildId]
-      build.status = 'running'
-      build.updated = now
-      if (!('start' in build)) {
-        build.start = now
-      }
-
-      let { steps } = build
-      let step = steps.find((s) => s.description === description)
-      if (!step) {
-        step = { description, start: now }
-        steps.push(step)
-      }
-      step.status = status
-      step.output = output
-
-      if (status === 'success' || status === 'failed') {
-        step.end = now
-        step.duration = step.end - step.start
-      }
-    })
-  },
-
-  async finishBuild(buildId, buildStatus) {
-    let [repoName] = buildId.split('#')
-
-    await updateRepoStatus(repoName, (repoStatus, now) => {
-      let build = repoStatus.builds[buildId]
-
-      build.status = buildStatus
-      build.end = now
-      build.updated = now
-      build.duration = build.end - build.start
-    })
-  }
+  logger.debug('finished updating status pages', { module: 'status/render' })
 }
