+++ lib/utils/misc.js
@@ -1,7 +1,8 @@
 const Git = require('nodegit')
+
 const {
   git: { authMethod, privateKey, publicKey, keyPassword }
-} = require('./config')
+} = require('../config')
 
 module.exports = {
   gitFetchOpts: {
