+++ lib/utils/queue.js
@@ -1,0 +1,15 @@
+class Queue {
+  constructor() {
+    this.promise = Promise.resolve()
+  }
+
+  run(asyncjob) {
+    this.promise = this.promise.then(asyncjob)
+  }
+
+  async join() {
+    await this.promise
+  }
+}
+
+module.exports = Queue
