+++ lib/watch/watcher.js
@@ -8,18 +8,19 @@ const { resolve } = require('path')
 const {
   workingDirectory,
   watcher: { interval }
-} = require('./config')
-const logger = require('./logger')
-const { gitFetchOpts } = require('./utils')
+} = require('../config')
+const logger = require('../utils/logger')('watcher')
+const { gitFetchOpts } = require('../utils/misc')
+
 const reposDirectory = resolve(workingDirectory, 'repos')
 
 class Watcher extends EventEmitter {
   constructor(repoName, repoUrl, branches) {
     super()
 
-    logger.info(`creating watcher for ${repoUrl} on ${branches.join(', ')}`, {
-      module: `watcher/${repoName}`
-    })
+    this.info(
+      `creating watcher for ${repoUrl} on branches ${branches.join(', ')}`
+    )
 
     this.repoName = repoName
     this.repoUrl = repoUrl
@@ -28,10 +29,18 @@ class Watcher extends EventEmitter {
     this.checkUpdates()
   }
 
+  debug(msg) {
+    logger.debug(msg, { module: `watcher/${this.repoName}` })
+  }
+
+  info(msg) {
+    logger.info(msg, { module: `watcher/${this.repoName}` })
+  }
+
   async checkUpdates() {
     try {
       let repo
-      let { repoName, repoUrl, branches } = this
+      let { repoUrl, branches } = this
       let cloned = false
       let repoPath = resolve(reposDirectory, this.repoName)
 
@@ -45,15 +54,11 @@ class Watcher extends EventEmitter {
 
       try {
         repo = await Git.Repository.open(repoPath)
-        logger.debug(`opened repo from ${repoPath}`, {
-          module: `watcher/${repoName}`
-        })
+        this.debug(`opened repo from ${repoPath}`)
       } catch(e) {
         // eslint-disable-next-line new-cap
         repo = await Git.Clone(repoUrl, repoPath, { fetchOpts: gitFetchOpts })
-        logger.debug(`cloned repo into ${repoPath}`, {
-          module: `watcher/${repoName}`
-        })
+        this.debug(`cloned repo into ${repoPath}`)
         cloned = true
       }
 
@@ -63,9 +68,9 @@ class Watcher extends EventEmitter {
           currentSHAs[branch] = (await repo.getBranchCommit(
             `origin/${branch}`
           )).sha()
-          logger.debug(`current SHA for ${branch} is ${currentSHAs[branch]}`, {
-            module: `watcher/${repoName}`
-          })
+          this.debug(
+            `current SHA for branch ${branch} is ${currentSHAs[branch]}`
+          )
         }
 
         await repo.fetch('origin', gitFetchOpts)
@@ -73,14 +78,10 @@ class Watcher extends EventEmitter {
 
       for (let branch of branches) {
         let newSHA = (await repo.getBranchCommit(`origin/${branch}`)).sha()
-        logger.debug(`updated SHA for ${branch} is ${newSHA}`, {
-          module: `watcher/${repoName}`
-        })
+        this.debug(`updated SHA for branch ${branch} is ${newSHA}`)
 
         if (newSHA !== currentSHAs[branch]) {
-          logger.info(`${branch} changed, new SHA is ${newSHA}`, {
-            module: `watcher/${repoName}`
-          })
+          this.info(`branch ${branch} changed, new SHA is ${newSHA}`)
           this.emit('change', branch, newSHA)
         }
       }
