+++ lib/watch/webhooks.js
@@ -4,8 +4,8 @@ const GithubWebHook = require('express-github-webhook')
 
 const {
   webhooks: { port, secret }
-} = require('./config')
-const logger = require('./logger')
+} = require('../config')
+const logger = require('../utils/logger')('webhooks')
 
 module.exports = function() {
   // eslint-disable-next-line new-cap
