+++ package.json
@@ -22,6 +22,7 @@
     "js-yaml": "^3.12.1",
     "nodegit": "^0.24.0",
     "tar": "^4.4.8",
+    "rsync": "^0.6.1",
     "winston": "^3.1.0"
   }
 }
